Create docker settings and configuration files for Showcase project
