+++ Smart.FA.Catalog.sln
@@ -19,6 +19,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		startup.bat = startup.bat
 		docker_startup_script.bat = docker_startup_script.bat
 		README.md = README.md
+		Showcase.Dockerfile = Showcase.Dockerfile
 	EndProjectSection
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "test", "test", "{D485BE00-A082-47D7-827D-0FD646807006}"
