+++ launch_useradmin.sh
