+++ src/Shared/Smart.FA.Catalog.Shared/Smart.FA.Catalog.Shared.csproj
@@ -10,4 +10,17 @@
     <ProjectReference Include="..\..\UserAdmin\src\Smart.FA.Catalog.Localization\Smart.FA.Catalog.Localization.csproj" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Reference Include="Microsoft.Extensions.Logging.Abstractions">
+      <HintPath>..\..\..\..\..\..\..\..\Program Files\dotnet\shared\Microsoft.AspNetCore.App\6.0.4\Microsoft.Extensions.Logging.Abstractions.dll</HintPath>
+    </Reference>
+    <Reference Include="Smart.FA.Catalog.Infrastructure">
+      <HintPath>..\..\UserAdmin\src\Smart.FA.Catalog.Web\bin\Debug\net6.0\Smart.FA.Catalog.Infrastructure.dll</HintPath>
+    </Reference>
+  </ItemGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.Extensions.Logging" Version="6.0.0" />
+  </ItemGroup>
+
 </Project>
