+++ src/Showcase/src/Smart.FA.Catalog.Showcase.Web/Program.cs
@@ -1,8 +1,13 @@
 using Microsoft.EntityFrameworkCore;
+using NLog.Web;
+using Smart.FA.Catalog.Shared.Security;
 using Smart.FA.Catalog.Showcase.Infrastructure.Data;
 using Smart.FA.Catalog.Showcase.Web.Extensions;
 
 var builder = WebApplication.CreateBuilder(args);
+
+// Use NLog as logging provider
+builder.Host.UseNLog();
 
 // Add services to the container.
 builder.Services.AddRazorPages();
@@ -15,7 +20,12 @@ builder.Services.AddDbContext<CatalogShowcaseContext>((_, efOptions) =>
 });
 builder.Services.AddHttpContextAccessor();
 
+
 var app = builder.Build();
+
+// A fix which allows SQL domain authentication with Kerberos
+NetSecurityNativeFix.Initialize(app.Services.GetRequiredService<ILogger<Program>>());
+
 
 app.UseStatusCodePagesWithReExecute("/{0}");
