+++ src/Showcase/src/Smart.FA.Catalog.Showcase.Web/appsettings.PreProduction.json
@@ -7,6 +7,6 @@
     }
   },
   "ConnectionStrings": {
-    "Catalog": "Server=host.docker.internal,1433; Database=Catalog; Integrated Security=true;"
+    "Catalog": "Server=SRVWSTGSQL\\SQLSTG;Database=SMart;Integrated Security=True;TrustServerCertificate=true"
   }
 }
